Move banner into notification.tsx
